// projects/facility-management/src/app/bookit/models/bookit.enum.ts
/** enums for repeat by options */
export enum BookItRepeatByOption {
    DAY_OF_MONTH = 1,
    DAY_OF_WEEK = 2
}

/** BookItRepeatsOnOption */
export enum BookItRepeatsOnOption {
    DAILY = 1,
    WEEKLY = 2,
    BIWEEKLY = 3,
    MONTHLY = 4,
    YEARLY = 5
}

/** navigate option */
export enum Navigate_Option {
    NAVIGATE_TO_ADD = '../',
    NAVIGATE_TO_EDIT = '../../',
}